Small changes on file imports to reflect move of the package
