+++ python/lsst/pipe/supertask/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import absolute_import
+from .activator import *
+from .super_task import *
