+++ python/lsst/pipe/supertask/examples/test2task.py
@@ -3,8 +3,8 @@ Test2 Task
 """
 from __future__ import absolute_import, division, print_function
 from lsst.pipe.base.task import Task
-from lsst.pipe.base.super_task import SuperTask
-import lsst.pipe.base.super_task as stask
+from lsst.pipe.supertask.super_task import SuperTask
+import lsst.pipe.supertask.super_task as stask
 from lsst.pipe.base.struct import Struct
 import lsst.pex.config as pexConfig
